integrate infomative_class
integrate informative_class interface among store component (only preparation, no implementation)
