+++ code/Sadna-17-B/Sadna-17-B.csproj
@@ -175,10 +175,12 @@
     <Compile Include="DomainLayer\Order\Order.cs" />
     <Compile Include="DomainLayer\Order\OrderSystem.cs" />
     <Compile Include="DomainLayer\Order\SubOrder.cs" />
-    <Compile Include="DomainLayer\Store\DiscountPolicy.cs" />
+    <Compile Include="DomainLayer\Store\Policies\Discount.cs" />
+    <Compile Include="DomainLayer\Store\Policies\DiscountPolicy.cs" />
     <Compile Include="DomainLayer\Store\Inventory.cs" />
+    <Compile Include="DomainLayer\Store\Policies\DiscountStrategy.cs" />
     <Compile Include="DomainLayer\Store\Product.cs" />
-    <Compile Include="DomainLayer\Store\PurchasePolicy.cs" />
+    <Compile Include="DomainLayer\Store\Policies\PurchasePolicy.cs" />
     <Compile Include="DomainLayer\Store\Store.cs" />
     <Compile Include="DomainLayer\Store\StoreController.cs" />
     <Compile Include="DomainLayer\User\Admin.cs" />
@@ -211,6 +213,7 @@
     <Compile Include="ServiceLayer\Services\StoreService.cs" />
     <Compile Include="ServiceLayer\Services\UserService.cs" />
     <Compile Include="Utils\Cryptography.cs" />
+    <Compile Include="Utils\informative.cs" />
     <Compile Include="Utils\Logger.cs" />
     <Compile Include="Utils\Response.cs" />
     <Compile Include="Utils\Sadna17BException.cs" />
